feat: state management
